Small fix to SMB_LIBRARY macro
It seems, that SMB_LIBRARY macro has small bug in the logic, when showing
if shall the SHARED version of the library be build.

If the default value is given as a parameter, it reports "yes" when library
is going to be build(?).

This small patch makes report consistent.

With regards,
Timur Bakeyev.

Signed-off-by: Stefan Metzmacher <metze@samba.org>