s4:torture Make RPC-NETLOGON pass against ncaclrpc servers
The original patch didn't cope with a NULL target server name - we now key off that to decide it isn't worth checking against LDAP for this host.

I still can't get this to pass against Windows 2008, but mdw was
testing against Windows 2008R2.  at least 'make test' is happy, and
the rest should not be too hard...

Andrew Bartlett