wbinfo: Reduce test noise in the wbinfo blackbox test
With the switch to libwbclient the previously stubbed out
--trusted-domains and --all-domains calls now fail. Set them to knownfail.

The previously knownfail -D test is now stubbed out, test it now.

This does not fix the issues with wbinfo -a and wbinfo -K not working on the
build farm. I have no idea whatsoever what is causing this, as those are broken
on my local machine even without my changes.